LayoutTests/fast/parser/element-nesting-cap-expected.txt

Test that the HTML parser does not allow the nesting depth of any elements to exceed MAX_DOM_TREE_DEPTH.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS s3.parentNode === s1 is true
PASS s2.parentNode === s1 is true
PASS s3.previousSibling === s2 is true
PASS successfullyParsed is true

TEST COMPLETE
